# openapi.json
{
    "openapi": "3.1.0",
    "info": {
        "title": "OpenAPI definition",
        "version": "v0"
    },
    "servers": [
        {
            "url": "http://localhost:8080",
            "description": "Generated server url"
        }
    ],
    "paths": {
...
        "/api/ai/seasons": {
            "post": {
                "tags": [
                    "ai-season-controller"
                ],
                "operationId": "createSeason",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CreateSeasonHttpRequest"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiResponseAiSeasonStatusView"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/ai/seasons/from-survey": {
            "post": {
                "tags": [
                    "ai-season-controller"
                ],
                "operationId": "createSeasonFromSurvey",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CreateSeasonFromSurveyRequest"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiResponseAiSeasonStatusView"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/ai/seasons/episodes/{number}/seen": {
            "post": {
                "tags": [
                    "ai-season-controller"
                ],
                "operationId": "markMessagesSeen",
                "parameters": [
                    {
                        "name": "number",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int32"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/MarkSeenHttpRequest"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiResponseVoid"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/ai/seasons/episodes/{number}/rerolls": {
            "get": {
                "tags": [
                    "ai-season-controller"
                ],
                "operationId": "getRerollStatus",
                "parameters": [
                    {
                        "name": "number",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int32"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiResponseAiRerollView"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "ai-season-controller"
                ],
                "operationId": "requestReroll",
                "parameters": [
                    {
                        "name": "number",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int32"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/RerollHttpRequest"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiResponseAiRerollView"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/ai/seasons/episodes/{number}/generations": {
            "post": {
                "tags": [
                    "ai-season-controller"
                ],
                "operationId": "requestGeneration",
                "parameters": [
                    {
                        "name": "number",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int32"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiResponseAiEpisodeProgressView"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/ai/seasons/episodes/3/selection": {
            "post": {
                "tags": [
                    "ai-season-controller"
                ],
                "operationId": "submitEpisode3Result",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Episode3ResultHttpRequest"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiResponseAiSelectionView"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/ai/seasons/episodes/1/selection": {
            "post": {
                "tags": [
                    "ai-season-controller"
                ],
                "operationId": "submitEpisode1Selection",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Episode1SelectionHttpRequest"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiResponseAiSelectionView"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/api/episodes/{episodeId}": {
            "get": {
                "tags": [
                    "episode-controller"
                ],
                "operationId": "getEpisode",
                "parameters": [
                    {
                        "name": "episodeId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/EpisodeResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/api/ai/seasons/episodes/{number}": {
            "get": {
                "tags": [
                    "ai-season-controller"
                ],
                "operationId": "getEpisodeMessages",
                "parameters": [
                    {
                        "name": "number",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int32"
                        }
                    },
                    {
                        "name": "after_sequence",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "format": "int32"
                        }
                    },
                    {
                        "name": "limit",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "format": "int32"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiResponseAiEpisodeMessagesView"
                                }
                            }
                        }
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
...
            "EpisodeResponse": {
                "type": "object",
                "properties": {
                    "episodeId": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "episodeNumber": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "title": {
                        "type": "string"
                    },
                    "status": {
                        "type": "string"
                    }
                }
            },
...
            "ApiResponseVoid": {
                "type": "object",
                "properties": {
                    "success": {
                        "type": "boolean"
                    },
                    "data": {
                        "type": "object"
                    },
                    "message": {
                        "type": "string"
                    },
                    "code": {
                        "type": "string"
                    }
                }
            },
...
            "AiCharacterProfileRequest": {
                "type": "object",
                "properties": {
                    "gender": {
                        "type": "string"
                    },
                    "traits": {
                        "$ref": "#/components/schemas/AiTraitsRequest"
                    }
                }
            },
            "AiTraitsRequest": {
                "type": "object",
                "properties": {
                    "affectionExpression": {
                        "type": "number"
                    },
                    "relationshipAnxiety": {
                        "type": "number"
                    },
                    "relationshipAvoidance": {
                        "type": "number"
                    },
                    "emotionalAttunement": {
                        "type": "number"
                    },
                    "relationshipInitiative": {
                        "type": "number"
                    },
                    "practicalPriority": {
                        "type": "number"
                    },
                    "reassuranceNeed": {
                        "type": "number"
                    },
                    "jealousyReactivity": {
                        "type": "number"
                    },
                    "relationshipEnergyDependence": {
                        "type": "number"
                    },
                    "emotionalSuppression": {
                        "type": "number"
                    },
                    "conflictConfrontation": {
                        "type": "number"
                    },
                    "relationshipPace": {
                        "type": "number"
                    },
                    "interestExpressionFrequency": {
                        "type": "number"
                    }
                }
            },
            "CreateSeasonHttpRequest": {
                "type": "object",
                "properties": {
                    "representative": {
                        "$ref": "#/components/schemas/AiCharacterProfileRequest"
                    }
                },
                "required": [
                    "representative"
                ]
            },
            "AiCharacterView": {
                "type": "object",
                "properties": {
                    "characterId": {
                        "type": "string",
                        "format": "uuid"
                    },
                    "role": {
                        "type": "string"
                    },
                    "gender": {
                        "type": "string"
                    }
                }
            },
            "AiSeasonStatusView": {
                "type": "object",
                "properties": {
                    "seasonId": {
                        "type": "string",
                        "format": "uuid"
                    },
                    "revision": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "currentEpisode": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "status": {
                        "type": "string"
                    },
                    "characters": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/AiCharacterView"
                        }
                    },
                    "sourceSurveyResultId": {
                        "type": "string",
                        "format": "uuid"
                    }
                }
            },
            "ApiResponseAiSeasonStatusView": {
                "type": "object",
                "properties": {
                    "success": {
                        "type": "boolean"
                    },
                    "data": {
                        "$ref": "#/components/schemas/AiSeasonStatusView"
                    },
                    "message": {
                        "type": "string"
                    },
                    "code": {
                        "type": "string"
                    }
                }
            },
            "CreateSeasonFromSurveyRequest": {
                "type": "object",
                "properties": {
                    "surveyResultId": {
                        "type": "string",
                        "format": "uuid"
                    }
                },
                "required": [
                    "surveyResultId"
                ]
            },
            "MarkSeenHttpRequest": {
                "type": "object",
                "properties": {
                    "sequence": {
                        "type": "integer",
                        "format": "int32",
                        "minimum": 0
                    }
                }
            },
            "RerollHttpRequest": {
                "type": "object",
                "properties": {
                    "partnerId": {
                        "type": "string",
                        "format": "uuid"
                    }
                },
                "required": [
                    "partnerId"
                ]
            },
            "AiRerollView": {
                "type": "object",
                "properties": {
                    "jobId": {
                        "type": "string",
                        "format": "uuid"
                    },
                    "jobStatus": {
                        "type": "string"
                    },
                    "errorCode": {
                        "type": "string"
                    }
                }
            },
            "ApiResponseAiRerollView": {
                "type": "object",
                "properties": {
                    "success": {
                        "type": "boolean"
                    },
                    "data": {
                        "$ref": "#/components/schemas/AiRerollView"
                    },
                    "message": {
                        "type": "string"
                    },
                    "code": {
                        "type": "string"
                    }
                }
            },
            "AiEpisodeProgressView": {
                "type": "object",
                "properties": {
                    "episodeNumber": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "jobId": {
                        "type": "string",
                        "format": "uuid"
                    },
                    "jobStatus": {
                        "type": "string"
                    },
                    "versionId": {
                        "type": "string",
                        "format": "uuid"
                    },
                    "errorCode": {
                        "type": "string"
                    }
                }
            },
            "ApiResponseAiEpisodeProgressView": {
                "type": "object",
                "properties": {
                    "success": {
                        "type": "boolean"
                    },
                    "data": {
                        "$ref": "#/components/schemas/AiEpisodeProgressView"
                    },
                    "message": {
                        "type": "string"
                    },
                    "code": {
                        "type": "string"
                    }
                }
            },
            "Episode3ResultHttpRequest": {
                "type": "object",
                "properties": {
                    "success": {
                        "type": "boolean"
                    },
                    "partnerId": {
                        "type": "string",
                        "format": "uuid"
                    }
                }
            },
            "AiSelectionView": {
                "type": "object",
                "properties": {
                    "selectionId": {
                        "type": "string",
                        "format": "uuid"
                    },
                    "partnerId": {
                        "type": "string",
                        "format": "uuid"
                    },
                    "revision": {
                        "type": "integer",
                        "format": "int32"
                    }
                }
            },
            "ApiResponseAiSelectionView": {
                "type": "object",
                "properties": {
                    "success": {
                        "type": "boolean"
                    },
                    "data": {
                        "$ref": "#/components/schemas/AiSelectionView"
                    },
                    "message": {
                        "type": "string"
                    },
                    "code": {
                        "type": "string"
                    }
                }
            },
            "Episode1SelectionHttpRequest": {
                "type": "object",
                "properties": {
                    "partnerId": {
                        "type": "string",
                        "format": "uuid"
                    }
                },
                "required": [
                    "partnerId"
                ]
            },
...
            "AiEpisodeMessagesView": {
                "type": "object",
                "properties": {
                    "episodeNumber": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "versionId": {
                        "type": "string",
                        "format": "uuid"
                    },
                    "topic": {
                        "$ref": "#/components/schemas/AiTopicView"
                    },
                    "messages": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/AiMessageView"
                        }
                    },
                    "hasMore": {
                        "type": "boolean"
                    },
                    "nextAfterSequence": {
                        "type": "integer",
                        "format": "int32"
                    }
                }
            },
            "AiMessageView": {
                "type": "object",
                "properties": {
                    "messageId": {
                        "type": "string",
                        "format": "uuid"
                    },
                    "sequence": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "sceneKind": {
                        "type": "string",
                        "enum": [
                            "group",
                            "candidates_only",
                            "one_to_one"
                        ]
                    },
                    "speakerId": {
                        "type": "string",
                        "format": "uuid"
                    },
                    "fromRepresentative": {
                        "type": "boolean"
                    },
                    "text": {
                        "type": "string"
                    }
                }
            },
            "AiTopicView": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "string"
                    },
                    "title": {
                        "type": "string"
                    },
                    "category": {
                        "type": "string"
                    }
                }
            },
            "ApiResponseAiEpisodeMessagesView": {
                "type": "object",
                "properties": {
                    "success": {
                        "type": "boolean"
                    },
                    "data": {
                        "$ref": "#/components/schemas/AiEpisodeMessagesView"
                    },
                    "message": {
                        "type": "string"
                    },
                    "code": {
                        "type": "string"
                    }
                }
            }
        }
    }
}